Began Creation of Physical Network Diagram
Devices are named according to Department + Room + type of device + Number for endpoints As an example, the 1st workstation in the HR department would be named HR101-PC01. For networking devices the convention is Network Device Type + Room + Room Number. As an example, router 1 in the network closet is labeled R1-NC101.